+++ include/apex/sarc.h
@@ -27,21 +27,21 @@ struct SArcEntry {
 
 class SArchive : public Archive {
 public:
-    SArchive(uint32 m_hash, std::unique_ptr<IO::File> buffer);
+    SArchive(uint64 m_hash, std::unique_ptr<IO::File> buffer);
 
     [[nodiscard]] bool has_file(std::string_view path) override;
 
-    [[nodiscard]] bool has_file(uint32 hash) override;
+    [[nodiscard]] bool has_file(uint64 hash) override;
 
     std::unique_ptr<IO::File> get_file(std::string_view path) override;
 
-    std::unique_ptr<IO::File> get_file(uint32 hash) override;
+    std::unique_ptr<IO::File> get_file(uint64 hash) override;
 
     void all_entries(std::vector<ArchiveEntry> &entries) const override;
 
     [[nodiscard]] std::string get_name() const override;
 
-    uint32 hash() override;
+    uint64 hash() override;
 
     [[nodiscard]] auto entries() const {
         return m_entries|std::views::values;
@@ -49,9 +49,9 @@ public:
 
 private:
     SArcHeader m_header{};
-    uint32 m_hash;
+    uint64 m_hash;
     std::vector<char> m_strings;
-    std::unordered_map<uint32, SArcEntry> m_entries;
+    std::unordered_map<uint64, SArcEntry> m_entries;
     std::unique_ptr<IO::File> m_buffer;
 };
